Inline SendError definition
